requirements.txt
# Dependências Python do projeto ARPIA
Django>=4.2,<5
djangorestframework==3.16.1
python-dotenv>=1.0
psycopg2-binary>=2.9
prometheus-client>=0.16
python-gvm>=23.0.0
pyattck>=7.1

# Imagens, static serving and template helpers
Pillow>=10.0
whitenoise>=6.5
django-widget-tweaks>=1.4.12
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7

# Testes
pytest>=7.0
pytest-django>=4.5

# Dependências de sistema (instale manualmente quando necessário):
#   jq .............. sudo apt install -y jq
#   rustscan ........ sudo apt update && sudo apt install -y cargo && cargo install rustscan